requirements.txt
google
discord.py
elasticsearch
